[FEAT] 슬랙 로그인 이름 형식 에러 처리

+++ FE/src/apis/instance.ts
@@ -63,8 +63,17 @@ https.interceptors.response.use(
       return await axios(originalRequest);
     }
 
+    if (errorCode === ERROR_CODE.AUTH.INVALID_NAME) {
+      toast.error(errorMessage, {
+        toastId: errorCode,
+      });
+      return Promise.reject(error);
+    }
+
     if (Object.values(ERROR_CODE.AUTH).includes(errorCode)) {
-      toast.error(errorMessage);
+      toast.error(errorMessage, {
+        toastId: errorCode,
+      });
       deleteTokenInfo();
       setTimeout(() => {
         window.location.href = "/login";
